create the query to get the list of chat-friends list

+++ my-app/backend/src/resolvers/index.js
@@ -1,5 +1,5 @@
 // File: backend/src/resolvers/index.js
-import bcrypt from "bcrypt";
+import bcrypt, { compareSync } from "bcrypt";
 import jwt from "jsonwebtoken";
 import {
   AuthenticationError,
@@ -29,7 +29,8 @@ const resolvers = {
     // Query work like GET
     hello: () => "Hello World! 🌍",
     getNotification: async (_, __, context) => {
-      const userId = isAuthorizedUserRequest(context);
+      const userId = getUserIdFromToken(context);
+      if (!userId) throw new AuthenticationError("Unauthorized");
 
       return await prisma.notification.findMany({
         where: { toUserId: userId },
@@ -37,7 +38,8 @@ const resolvers = {
       });
     },
     getpost: async (_, { id }, context) => {
-      const userId = isAuthorizedUserRequest(context);
+      const userId = getUserIdFromToken(context);
+      if (!userId) throw new AuthenticationError("Unauthorized");
 
       const post = await prisma.post.findUnique({
         where: {
@@ -62,6 +64,28 @@ const resolvers = {
         token,
         mycomment,
       };
+    },
+
+    friendChatList: async (_, { userId }, context) => {
+      const token = getToken(context);
+      if (!token) throw new AuthenticationError("Unauthorized");
+
+      const chatRoomList = await prisma.participant.findMany({
+        where: { userId },
+        include: {
+          user: true,
+          chatRoom: true,
+        },
+      });
+
+      console.log(
+        "Get the List of ChatroomId is Array ",
+        Array.isArray(chatRoomList)
+      );
+
+      console.log("List: ", chatRoomList);
+
+      return chatRoomList;
     },
   },
   Mutation: {
@@ -141,7 +165,8 @@ const resolvers = {
       };
     },
     createpost: async (_, { title, content, category }, context) => {
-      const userId = isAuthorizedUserRequest(context);
+      const userId = getUserIdFromToken(context);
+      if (!userId) throw new AuthenticationError("Unauthorized");
 
       const post = await prisma.post.create({
         data: {
@@ -164,7 +189,8 @@ const resolvers = {
       };
     },
     updatepost: async (_, { id, title, content, category }, context) => {
-      const userId = isAuthorizedUserRequest(context);
+      const userId = getUserIdFromToken(context);
+      if (!userId) throw new AuthenticationError("Unauthorized");
 
       const data = {};
       if (title) {
@@ -187,7 +213,8 @@ const resolvers = {
       };
     },
     deletepost: async (_, { id }, context) => {
-      const userId = isAuthorizedUserRequest(context);
+      const userId = getUserIdFromToken(context);
+      if (!userId) throw new AuthenticationError("Unauthorized");
 
       const post = await prisma.post.delete({
         where: { id },
@@ -198,7 +225,8 @@ const resolvers = {
       };
     },
     likepost: async (_, { postId }, context) => {
-      const userId = isAuthorizedUserRequest(context);
+      const userId = getUserIdFromToken(context);
+      if (!userId) throw new AuthenticationError("Unauthorized");
 
       const token = getToken(context);
       if (!token) throw new AuthenticationError("Priviledged violation!");
@@ -262,7 +290,8 @@ const resolvers = {
       }
     },
     updateprofile: async (_, { name, email }, context) => {
-      const userId = isAuthorizedUserRequest(context);
+      const userId = getUserIdFromToken(context);
+      if (!userId) throw new AuthenticationError("Unauthorized");
 
       const data = {};
 
@@ -288,7 +317,8 @@ const resolvers = {
       };
     },
     deleteuser: async (_, { id }, context) => {
-      const userId = isAuthorizedUserRequest(context);
+      const userId = getUserIdFromToken(context);
+      if (!userId) throw new AuthenticationError("Unauthorized");
 
       const isAdmin = checkIsAdmin(context);
       if (
@@ -309,7 +339,8 @@ const resolvers = {
       };
     },
     createcomment: async (_, { comment, postId }, context) => {
-      const authorId = isAuthorizedUserRequest(context);
+      const authorId = getUserIdFromToken(context);
+      if (!authorId) throw new AuthenticationError("Unauthorized");
 
       const token = getToken(context);
       if (!token) throw new AuthenticationError("Priviledged violation!");
@@ -324,7 +355,8 @@ const resolvers = {
       };
     },
     updatecomment: async (_, { comment, id }, context) => {
-      const authorId = isAuthorizedUserRequest(context);
+      const authorId = getUserIdFromToken(context);
+      if (!authorId) throw new AuthenticationError("Unauthorized");
 
       const token = getToken(context);
       if (!token) throw new AuthenticationError("Priveledged violation!");
@@ -344,7 +376,8 @@ const resolvers = {
       };
     },
     deletecomment: async (_, { id }, context) => {
-      const authorId = isAuthorizedUserRequest(context);
+      const authorId = getUserIdFromToken(context);
+      if (!authorId) throw new AuthenticationError("Unauthorized");
 
       const token = getToken(context);
       if (!token) throw new AuthenticationError("Priveledged violation!");
@@ -358,7 +391,8 @@ const resolvers = {
       };
     },
     likecomment: async (_, { postId, commentId }, context) => {
-      const userId = isAuthorizedUserRequest(context);
+      const userId = getUserIdFromToken(context);
+      if (!userId) throw new AuthenticationError("Unauthorized");
 
       try {
         const existing = await prisma.commentLike.findUnique({
@@ -419,7 +453,8 @@ const resolvers = {
       }
     },
     friendSendRequest: async (_, { receiverId }, context) => {
-      const senderId = isAuthorizedUserRequest(context);
+      const senderId = getUserIdFromToken(context);
+      if (!senderId) throw new AuthenticationError("Unauthorized");
 
       const req = await prisma.friendRequest.create({
         data: { senderId, receiverId, type: "FRIEND_REQUEST_SENT" },
@@ -495,7 +530,8 @@ const resolvers = {
     },
 
     activateChatRoom: async (_, { targetUserId }, context) => {
-      const userId = isAuthorizedUserRequest(context);
+      const userId = getUserIdFromToken(context);
+      if (!userId) throw new AuthenticationError("Unauthorized");
 
       const isChatRoomExist = await prisma.chatRoom.findFirst({
         where: {
@@ -575,7 +611,8 @@ const resolvers = {
       };
     },
     textMessage: async (_, { chatRoomId, content }, context) => {
-      const userId = isAuthorizedUserRequest(context);
+      const userId = getUserIdFromToken(context);
+      if (!userId) throw new AuthenticationError("Unauthorized");
 
       const msg = await prisma.message.create({
         data: {
@@ -600,7 +637,8 @@ const resolvers = {
       };
     },
     iNotify: async (_, { id }, context) => {
-      const userId = isAuthorizedUserRequest(context);
+      const userId = getUserIdFromToken(context);
+      if (!userId) throw new AuthenticationError("Unauthorized");
 
       const notify = await prisma.notification.findUnique({
         where: { id },
@@ -680,13 +718,6 @@ async function createNotification({
   });
 
   return notify;
-}
-
-function isAuthorizedUserRequest(context) {
-  const userId = getUserIdFromToken(context);
-  if (!userId) throw new AuthenticationError("Unauthorized");
-
-  return userId;
 }
 
 async function updateNotification({
